AccountFlags -> enum
